+++ persistent/Database/Persist/TH.hs
@@ -119,6 +119,7 @@ import Language.Haskell.TH.Syntax
 import Web.HttpApiData (FromHttpApiData(..), ToHttpApiData(..))
 import Web.PathPieces (PathPiece(..))
 
+import Database.Persist.Class.PersistEntity
 import Database.Persist
 import Database.Persist.Quasi
 import Database.Persist.Quasi.Internal
@@ -796,6 +797,7 @@ mkPersistWith mps preexistingEntities ents' = do
     entityDecs <- fmap mconcat $ mapM (mkEntity embedEntityMap entityMap mps) ents
     jsonDecs <- fmap mconcat $ mapM (mkJSON mps) ents
     uniqueKeyInstances <- fmap mconcat $ mapM (mkUniqueKeyInstances mps) ents
+    safeToInsertInstances <- mconcat <$> mapM (mkSafeToInsertInstance mps) ents
     symbolToFieldInstances <- fmap mconcat $ mapM (mkSymbolToFieldInstances mps entityMap) ents
     return $ mconcat
         [ persistFieldDecs
@@ -803,7 +805,54 @@ mkPersistWith mps preexistingEntities ents' = do
         , jsonDecs
         , uniqueKeyInstances
         , symbolToFieldInstances
+        , safeToInsertInstances
         ]
+
+mkSafeToInsertInstance :: MkPersistSettings -> UnboundEntityDef -> Q [Dec]
+mkSafeToInsertInstance mps ued =
+    case unboundPrimarySpec ued of
+        NaturalKey _ ->
+            instanceOkay
+        SurrogateKey uidDef -> do
+            let attrs =
+                    unboundIdAttrs uidDef
+                isDefaultFieldAttr = \case
+                    FieldAttrDefault _ ->
+                        True
+                    _ ->
+                        False
+            case List.find isDefaultFieldAttr attrs of
+                Nothing ->
+                    badInstance
+                Just _ ->
+                    instanceOkay
+
+        DefaultKey _ ->
+            instanceOkay
+
+  where
+    typ :: Type
+    typ = genericDataType mps (getUnboundEntityNameHS ued) backendT
+
+    mkInstance merr =
+        InstanceD Nothing (maybe id (:) merr withPersistStoreWriteCxt) (ConT ''SafeToInsert `AppT` typ) []
+    instanceOkay =
+        pure
+            [ mkInstance Nothing
+            ]
+    badInstance = do
+        err <- [t| TypeError (SafeToInsertErrorMessage $(pure typ)) |]
+        pure
+            [ mkInstance (Just err)
+            ]
+
+    withPersistStoreWriteCxt =
+        if mpsGeneric mps
+            then
+                [ConT ''PersistStoreWrite `AppT` backendT]
+            else
+                []
+
 
 -- we can't just use 'isInstance' because TH throws an error
 shouldGenerateCode :: UnboundEntityDef -> Q Bool